Updated goal test, goal state and result code.
Changes to goal test required to ensure it was comparing the two strings insted of string to seqence of actions. Also deleted player position in the test string so it matched the goal string
Changes to result code to allow for the string to account for boxes which are moved off slots and boxes which are moved onto slots.

Continued to implement solver but only seems to return the first new nodes and never a solution. Will continue to look in what may be causeing issues