Bug fixed in Check_elem_sequence that did not catch when player moved into wall
As the state is not updated if a player moves into a wall the check never saw the player in the wall. Instead now I am checking if the cordinates are the same as the previous action. If so the player has tried to move into a wall or box that cannot be pushed